Refactor sessions
