+++ src/renderer/components/Sessions/SessionGrid.scss
