+++ src/renderer/components/Sessions/SessionGrid.tsx
