+++ src/renderer/components/Sessions/SessionRow.tsx
